backend/requirements.txt
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
appnope==0.1.4
argcomplete==3.6.2
asttokens==3.0.0
boto3==1.38.13
botocore==1.38.13
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cohere==5.15.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.2
google-auth==2.40.1
griffe==1.7.3
groq==0.24.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
logfire-api==3.15.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mcp==1.8.0
mdurl==0.1.2
mistralai==1.7.0
nest-asyncio==1.6.0
openai==1.78.0
opentelemetry-api==1.33.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-ai==0.1.11
pydantic-ai-slim==0.1.11
pydantic-evals==0.1.11
pydantic-graph==0.1.11
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.46.2
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0
annotated-types==0.7.0
anthropic==0.51.0
anyio==4.9.0
appnope==0.1.4
argcomplete==3.6.2
asttokens==3.0.0
boto3==1.38.13
botocore==1.38.13
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cohere==5.15.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.12
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.2
google-auth==2.40.1
griffe==1.7.3
groq==0.24.0
h11==0.16.0
hf-xet==1.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.1
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.9.0
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.7.2
logfire-api==3.15.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mcp==1.8.0
mdurl==0.1.2
mistralai==1.7.0
nest-asyncio==1.6.0
openai==1.78.0
opentelemetry-api==1.33.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-ai==0.1.11
pydantic-ai-slim==0.1.11
pydantic-evals==0.1.11
pydantic-graph==0.1.11
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.46.2
tokenizers==0.21.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0
